ml_service/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pandas
scikit-learn
opencv-python
Pillow
scipy
shapely
requests
python-multipart
